oracle-2/services/ecg/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
neurokit2==0.2.7
numpy>=1.24.0
pandas==2.2.2
Pillow>=10.0.0
slowapi>=0.1.8
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.9
requests>=2.31.0

